Move filesystem fixture
